Create Tuition Calculator
Ok so I stuggled with this code. The Prompt:
Suppose that the tuition for a university is $10,000 this year and increases 5% every year. In one year, the tuition will be $10,500. Write a program that displays the tuition in ten years and the total cost of four years’ worth of tuition after the tenth year.
The correct output for the final cost of four years should be  $70207.39. My output returns 70355.02 dollars. I will revisit this code later.